Update Importer module spec
